+++ lib/rails_template18f/generators/cloud_gov_options.rb
@@ -4,6 +4,7 @@ module RailsTemplate18f
   module Generators
     module CloudGovOptions
       extend ActiveSupport::Concern
+      include CloudGovParsing
 
       included do
         class_option :cg_org, desc: "cloud.gov organization name"
@@ -14,39 +15,18 @@ module RailsTemplate18f
       private
 
       def cloud_gov_organization
-        if options[:cg_org].present?
-          return options[:cg_org]
-        elsif terraform_dir_exists?
-          staging_main = file_content("terraform/staging/main.tf")
-          if (matches = staging_main.match(/cf_org_name\s+= "(?<org_name>.*)"/))
-            return matches[:org_name]
-          end
-        end
-        "TKTK-cloud.gov-org-name"
+        return options[:cg_org] if options[:cg_org].present?
+        super
       end
 
       def cloud_gov_staging_space
-        if options[:cg_staging].present?
-          return options[:cg_staging]
-        elsif terraform_dir_exists?
-          staging_main = file_content("terraform/staging/main.tf")
-          if (matches = staging_main.match(/cf_space_name\s+= "(?<space_name>.*)"/))
-            return matches[:space_name]
-          end
-        end
-        "staging"
+        return options[:cg_staging] if options[:cg_staging].present?
+        super
       end
 
       def cloud_gov_production_space
-        if options[:cg_prod].present?
-          return options[:cg_prod]
-        elsif terraform_dir_exists?
-          prod_main = file_content("terraform/production/main.tf")
-          if (matches = prod_main.match(/cf_space_name\s+= "(?<space_name>.*)"/))
-            return matches[:space_name]
-          end
-        end
-        "prod"
+        return options[:cg_prod] if options[:cg_prod].present?
+        super
       end
     end
   end